requirements.txt
google-cloud-compute==1.15.0
redis==5.0.1
psycopg2-binary==2.9.9
pyzmq==25.1.2
streamlit==1.31.0
pandas==2.2.0
python-dotenv==1.0.1
asyncpg==0.29.0
polars==0.20.10
numpy==1.26.4
httpx==0.27.0
